requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 25.3.1

absl-py=2.2.2=pypi_0
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.11.18=pypi_0
aiosignal=1.3.2=pypi_0
annotated-types=0.7.0=pypi_0
asttokens=3.0.0=pyhd8ed1ab_1
attrs=25.3.0=pypi_0
blas=1.0=mkl
blis=1.3.0=pypi_0
brotli-python=1.1.0=py312h275cf98_2
bzip2=1.0.8=h2466b09_7
ca-certificates=2025.4.26=h4c7d964_0
catalogue=2.0.10=pypi_0
certifi=2025.1.31=pyhd8ed1ab_0
cffi=1.17.1=py312h4389bb4_0
charset-normalizer=3.4.2=pyhd8ed1ab_0
click=8.1.8=pypi_0
cloudpathlib=0.21.0=pypi_0
colorama=0.4.6=pyhd8ed1ab_1
comm=0.2.2=pyhd8ed1ab_1
confection=0.1.5=pypi_0
contourpy=1.3.2=pypi_0
cpython=3.12.10=py312hd8ed1ab_0
cuda-cccl=12.9.27=0
cuda-cccl_win-64=12.9.27=0
cuda-cudart=12.1.105=0
cuda-cudart-dev=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-libraries-dev=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvrtc-dev=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.9.19=0
cuda-opencl-dev=12.9.19=0
cuda-profiler-api=12.9.19=0
cuda-runtime=12.1.0=0
cuda-version=12.9=3
cycler=0.12.1=pypi_0
cymem=2.0.11=pypi_0
debugpy=1.8.14=py312h275cf98_0
decorator=5.2.1=pyhd8ed1ab_0
en-core-web-lg=3.8.0=pypi_0
en-core-web-sm=3.8.0=pypi_0
exceptiongroup=1.2.2=pyhd8ed1ab_1
executing=2.2.0=pyhd8ed1ab_0
faster-coco-eval=1.6.5=pypi_0
filelock=3.18.0=pyhd8ed1ab_0
fonttools=4.57.0=pypi_0
freetype=2.13.3=h57928b3_1
frozenlist=1.6.0=pypi_0
fsspec=2025.3.2=pypi_0
grpcio=1.71.0=pypi_0
h2=4.2.0=pyhd8ed1ab_0
hpack=4.1.0=pyhd8ed1ab_0
hyperframe=6.1.0=pyhd8ed1ab_0
idna=3.10=pyhd8ed1ab_1
importlib-metadata=8.6.1=pyha770c72_0
intel-openmp=2025.1.0=h57928b3_980
ipykernel=6.29.5=pyh4bbf305_0
ipython=9.2.0=pyhca29cf9_0
ipython_pygments_lexers=1.1.1=pyhd8ed1ab_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.6=pyhd8ed1ab_0
joblib=1.5.0=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.7.2=pyh5737063_1
khronos-opencl-icd-loader=2024.10.24=h2466b09_1
kiwisolver=1.4.8=pypi_0
krb5=1.21.3=hdf4eb48_0
langcodes=3.5.0=pypi_0
language-data=1.3.0=pypi_0
lcms2=2.17=hbcf6048_0
lerc=4.0.0=h6470a55_1
libblas=3.9.0=1_h8933c1f_netlib
libcblas=3.9.0=12_hb3dda5d_netlib
libcublas=12.1.0.26=0
libcublas-dev=12.1.0.26=0
libcufft=11.0.2.4=0
libcufft-dev=11.0.2.4=0
libcurand=10.3.10.19=0
libcurand-dev=10.3.10.19=0
libcusolver=11.4.4.55=0
libcusolver-dev=11.4.4.55=0
libcusparse=12.0.2.55=0
libcusparse-dev=12.0.2.55=0
libdeflate=1.23=h76ddb4d_0
libexpat=2.7.0=he0c23c2_0
libffi=3.4.6=h537db12_1
libfreetype=2.13.3=h57928b3_1
libfreetype6=2.13.3=h0b5ce68_1
libhwloc=2.11.2=default_hc8275d1_1000
libiconv=1.18=h135ad9c_1
libjpeg-turbo=3.1.0=h2466b09_0
liblapack=3.9.0=12_h13b7882_netlib
liblzma=5.8.1=h2466b09_1
libnpp=12.0.2.50=0
libnpp-dev=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjitlink-dev=12.1.105=0
libnvjpeg=12.1.1.14=0
libnvjpeg-dev=12.1.1.14=0
libpng=1.6.47=h7a4582a_0
libsodium=1.0.20=hc70643c_0
libsqlite=3.49.1=h67fdade_2
libtiff=4.7.0=h797046b_4
libuv=1.50.0=h2466b09_0
libwebp=1.5.0=h3b0e114_0
libwebp-base=1.5.0=h3b0e114_0
libxcb=1.16=h013a479_1
libxml2=2.13.7=h442d1da_1
libzlib=1.3.1=h2466b09_2
lightning=2.5.1.post0=pypi_0
lightning-utilities=0.14.3=pypi_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
marisa-trie=1.2.1=pypi_0
markdown=3.8=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=py312h31fea79_1
matplotlib-base=3.10.1=py312h90004f6_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h6a75c08_48682
mpmath=1.3.0=pyhd8ed1ab_1
msys2-conda-epoch=20160418=1
multidict=6.4.3=pypi_0
murmurhash=1.0.12=pypi_0
narwhals=1.38.0=pypi_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.4.2=pyh267e887_2
nltk=3.9.1=pypi_0
numpy=2.2.5=py312h3150e54_0
opencl-headers=2024.10.24=he0c23c2_0
opencv-python=4.11.0.86=pypi_0
openjpeg=2.5.3=h4d64b90_0
openssl=3.5.0=ha4e3fda_1
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
parso=0.8.4=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=10.4.0=py312h381445a_1
pip=25.1.1=pyh8b19718_0
platformdirs=4.3.7=pyh29332c3_0
plotly=6.0.1=pypi_0
preshed=3.0.9=pypi_0
prompt-toolkit=3.0.51=pyha770c72_0
propcache=0.3.1=pypi_0
protobuf=6.30.2=pypi_0
psutil=7.0.0=py312h4389bb4_0
pthread-stubs=0.4=hcd874cb_1001
pthreads-win32=2.9.1=h2466b09_4
pure_eval=0.2.3=pyhd8ed1ab_1
py-cpuinfo=9.0.0=pypi_0
pycocotools=2.0.8=py312h1a27103_2
pycparser=2.22=pyh29332c3_1
pydantic=2.11.4=pypi_0
pydantic-core=2.33.2=pypi_0
pygments=2.19.1=pyhd8ed1ab_0
pyparsing=3.2.3=pypi_0
pysocks=1.7.1=pyh09c184e_7
python=3.12.10=h3f84c4b_0_cpython
python-dateutil=2.9.0.post0=pyhff2d567_1
python_abi=3.12=7_cp312
pytorch=2.5.1=py3.12_cuda12.1_cudnn9_0
pytorch-cuda=12.1=hde6ce7c_6
pytorch-lightning=2.5.1.post0=pypi_0
pytorch-mutex=1.0=cuda
pytz=2025.2=pypi_0
pywin32=307=py312h275cf98_3
pyyaml=6.0.2=py312h31fea79_2
pyzmq=26.4.0=py312hd7027bb_0
qhull=2020.2=hc790b64_5
regex=2024.11.6=pypi_0
requests=2.32.3=pyhd8ed1ab_1
rich=14.0.0=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.15.2=pypi_0
seaborn=0.13.2=pypi_0
setuptools=80.1.0=pyhff2d567_0
shellingham=1.5.4=pypi_0
six=1.17.0=pyhd8ed1ab_0
smart-open=7.1.0=pypi_0
spacy=3.8.5=pypi_0
spacy-legacy=3.0.12=pypi_0
spacy-loggers=1.0.5=pypi_0
srsly=2.5.1=pypi_0
stack_data=0.6.3=pyhd8ed1ab_1
sympy=1.13.1=pypi_0
tbb=2021.13.0=h62715c5_1
tensorboard=2.19.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
thinc=8.3.6=pypi_0
threadpoolctl=3.6.0=pypi_0
tk=8.6.13=h5226925_1
torchaudio=2.5.1=pypi_0
torchinfo=1.8.0=pypi_0
torchmetrics=1.7.1=pypi_0
torchvision=0.20.1=pypi_0
tornado=6.4.2=py312h4389bb4_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pyhd8ed1ab_1
typer=0.15.3=pypi_0
typing-inspection=0.4.0=pypi_0
typing_extensions=4.13.2=pyh29332c3_0
tzdata=2025.2=pypi_0
ucrt=10.0.22621.0=h57928b3_1
ultralytics=8.3.127=pypi_0
ultralytics-thop=2.0.14=pypi_0
urllib3=2.4.0=pyhd8ed1ab_0
vc=14.3=h2b53caa_26
vc14_runtime=14.42.34438=hfd919c2_26
vs2015_runtime=14.42.34438=h7142326_26
wasabi=1.1.3=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_1
weasel=0.4.1=pypi_0
werkzeug=3.1.3=pypi_0
wget=3.2=pypi_0
wheel=0.45.1=pyhd8ed1ab_1
win_inet_pton=1.1.0=pyh7428d3b_8
wrapt=1.17.2=pypi_0
xorg-libxau=1.0.11=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
yaml=0.2.5=h8ffe710_2
yarl=1.20.0=pypi_0
zeromq=4.3.5=ha9f60a1_7
zipp=3.21.0=pyhd8ed1ab_1
zstandard=0.23.0=py312h4389bb4_2
zstd=1.5.7=hbeecb71_2
